refactor: move logic from professor controller to professor service

+++ src/main/java/com/luminabackend/services/ProfessorService.java
@@ -1,6 +1,10 @@
 package com.luminabackend.services;
 
+import com.luminabackend.exceptions.EmailAlreadyInUseException;
+import com.luminabackend.exceptions.EntityNotFoundException;
 import com.luminabackend.models.user.Professor;
+import com.luminabackend.models.user.User;
+import com.luminabackend.models.user.dto.user.UserPutDTO;
 import com.luminabackend.models.user.dto.user.UserSignupDTO;
 import com.luminabackend.repositories.professor.ProfessorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,22 +21,10 @@ public class ProfessorService {
     private ProfessorRepository repository;
 
     @Autowired
+    private AccountService accountService;
+
+    @Autowired
     private PasswordEncoder passwordEncoder;
-
-    public Professor save(UserSignupDTO professorPostDTO){
-        String email = professorPostDTO.email().trim();
-        String password = professorPostDTO.password().trim();
-        String encodedPassword = passwordEncoder.encode(password);
-        String firstName = professorPostDTO.firstName().trim();
-        String lastName = professorPostDTO.lastName().trim();
-
-        Professor professor = new Professor(email, encodedPassword, firstName, lastName);
-        return repository.save(professor);
-    }
-
-    public Professor save(Professor professor){
-        return repository.save(professor);
-    }
 
     public List<Professor> getAllProfessors() {
         return repository.findAll();
@@ -50,7 +42,33 @@ public class ProfessorService {
         return repository.existsById(id);
     }
 
+    public Professor save(UserSignupDTO professorPostDTO){
+        Optional<User> userByEmail = accountService.getUserByEmail(professorPostDTO.email());
+
+        if (userByEmail.isPresent()) throw new EmailAlreadyInUseException();
+
+        String email = professorPostDTO.email().trim();
+        String password = professorPostDTO.password().trim();
+        String encodedPassword = passwordEncoder.encode(password);
+        String firstName = professorPostDTO.firstName().trim();
+        String lastName = professorPostDTO.lastName().trim();
+
+        Professor professor = new Professor(email, encodedPassword, firstName, lastName);
+        return repository.save(professor);
+    }
+
+    public Professor edit(UUID id, UserPutDTO userPutDTO){
+        Optional<Professor> professorById = getProfessorById(id);
+        if(professorById.isEmpty()) throw new EntityNotFoundException("Professor not found");
+
+        Professor professor = professorById.get();
+        professor = (Professor) accountService.editUserData(professor, userPutDTO);
+        return repository.save(professor);
+    }
+
     public void deleteById(UUID id) {
+        if (!existsById(id)) throw new EntityNotFoundException("Professor not found");
+
         repository.deleteById(id);
     }
 }
